+++ go/trace/plugin_datadog.go
@@ -9,7 +9,7 @@ import (
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/opentracer"
 	ddtracer "gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 
-	"vitess.io/vitess/go/viperutil/v2"
+	"vitess.io/vitess/go/viperutil"
 )
 
 var (
