+++ go/viperutil/funcs/get_test.go
@@ -21,9 +21,9 @@ import (
 
 	"github.com/spf13/viper"
 
-	"vitess.io/vitess/go/viperutil/v2"
-	"vitess.io/vitess/go/viperutil/v2/funcs"
-	"vitess.io/vitess/go/viperutil/v2/internal/value"
+	"vitess.io/vitess/go/viperutil"
+	"vitess.io/vitess/go/viperutil/funcs"
+	"vitess.io/vitess/go/viperutil/internal/value"
 )
 
 func ExampleGetPath() {
