+++ go/viperutil/internal/registry/registry.go
@@ -20,7 +20,7 @@ import (
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 
-	"vitess.io/vitess/go/viperutil/v2/internal/sync"
+	"vitess.io/vitess/go/viperutil/internal/sync"
 )
 
 var (
